% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{fars_map_state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{The digit code assigned to each state}

\item{year}{Year for which data to be fetched as number or character}
}
\value{
A map plot with each dot representing the geospatial location where
accident happned within each state in that year.
}
\description{
Plots the map of the state based on state num and year
Each accident is a data point which is plotted at its geospatial coordinates
}
\note{
If state number is invalid i.e when the state code is not there in data file

When there are no accidents to plot, it displays "no accidents to plot"
}
\examples{
fars_map_state(49,2013)
fars_map_state("1","2015")
}
