% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{fars_read}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{Actual filename with extension as a character string}
}
\value{
A dataframe containing the full data
}
\description{
Second function to be called in fars_map_state function.
Loads the data as a dataframe from current working directory
If file doesn't exist in current directory, this throws a message "file <filename> does not exist"
}
\note{
Warning is given when file is not available in current working directory

The filename should be a character string
}
\examples{
data <- fars_read("accident_2013.csv.bz2")
}
