% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A character vector of years for which we want to subset data}
}
\value{
A list of dataframes where each list item has a dataframe for one year
The dataframe has two columns, MONTH and year
}
\description{
This function is used by fars_summarize_years for count summary on monthly basis
}
\note{
Any year outside the ones in datasets will throw an invalid year messsage
}
\examples{
fars_read_years(years = c(2013, 2014))
}
