% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A character vector of years for which we want count summary of fatality}
}
\value{
A summary dataframe with counts of fatality per month for years specified
}
\description{
Generates a count summary of fatality numbers per month for years specified as output
}
\note{
Error can happen when the year mentioned is not available in dataset
}
\examples{
fars_summarize_years(2013:2015)
fars_summarize_years(c(2013,2014))

}
