% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{4 digit year as character or number string}
}
\value{
A character string which is the filename
}
\description{
This is the first function to be called in fars_map_state function.
The function accepts year as input and returns a filename with bzip2 extension
}
\note{
Any character input will throw an error since input is a number

The function will throw a spanner later if we input say 20211 instead of 2021
}
\examples{
filename <- make_filename(2013)
filename <- make_filename("2015")
}
